+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+Changes:
+
+- Add missing barcode ID to export
+
 ## v1.14 (2021-06-07)
 
 Changes: